Attack Combo
Attack combo animation implemented through code.

Code changed in Chip class and ChipAnimInstance class.
If want to only be in chip class, a timer would need to determine when
the user can click to activate next combo. For now the Montage is setup
and called in the ChipAnimInstance Class.

Currently does not play through the combo anims